# engine/docs2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack API",
    "description": "Drop-in replacement for Algolia search. Provides full-text search with filters, facets, geo-search, and more.",
    "contact": {
      "name": "Flapjack",
      "url": "https://github.com/stuartcrobinson/flapjack"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7700",
      "description": "Local development"
    },
    {
      "url": "https://fj-us-west-1.flapjack.foo",
      "description": "Production US West"
    }
  ],
  "paths": {
...
    "/1/indexes/{indexName}/rules/batch": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Batch create or update multiple query rules in the specified index.",
        "operationId": "save_rules",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Array of rules",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Rules saved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid query parameter value"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/indexes/{indexName}/rules/clear": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Remove all query rules from the specified index.",
        "operationId": "clear_rules",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rules cleared",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/indexes/{indexName}/rules/search": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Search for query rules in the specified index with optional filtering.",
        "operationId": "search_rules",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Search parameters",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Matching rules",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/1/indexes/{indexName}/rules/{objectID}": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Retrieve a query rule by its object ID from the specified index.",
        "operationId": "get_rule",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "objectID",
            "in": "path",
            "description": "Rule ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rule retrieved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Rule not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "Save or update a single query rule in the specified index.",
        "operationId": "save_rule",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "objectID",
            "in": "path",
            "description": "Rule ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Rule data",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Rule saved",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "rules"
        ],
        "summary": "Delete a query rule from the specified index by its object ID.",
        "operationId": "delete_rule",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "objectID",
            "in": "path",
            "description": "Rule ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rule deleted",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Rule not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-algolia-api-key"
      }
    }
  },
...
}